Rename ActivityProfile folder for consistency
