+++ src/routes.js
@@ -1,7 +1,6 @@
 import About from './views/About';
 import InvolvementsAll from './views/InvolvementsAll';
-//import ActivityEdit from './views/ActivityEdit';
-import InvolvementProfile from './views/ActivityProfile';
+import InvolvementProfile from './views/InvolvementProfile';
 import Home from './views/Home';
 import WellnessCheck from './views/WellnessCheck';
 import ApartmentApp from './views/ApartmentApp';
